Add support codama device

+++ setup.sh
@@ -36,7 +36,9 @@ if [ -e /usr/share/alsa/pulse-alsa.conf ] ; then
 fi
 
 # Check args for asoundrc selection. Default to VF Stereo.
-if [ $# -eq 1 ] && [ $1 = "vocalfusion" ] ; then
+if [ $# -eq 1 ] && [ $1 = "codama" ] ; then
+    cp $RPI_SETUP_DIR/resources/asoundrc_vf_codama ~/.asoundrc
+elif [ $# -eq 1 ] && [ $1 = "vocalfusion" ] ; then
     cp $RPI_SETUP_DIR/resources/asoundrc_vf ~/.asoundrc
 elif [ $# -ge 1 ] && [ $1 = "xvf3510" ] ; then
     cp $RPI_SETUP_DIR/resources/asoundrc_vf_xvf3510 ~/.asoundrc
@@ -58,11 +60,16 @@ sudo /etc/init.d/alsa-utils restart
 # Create the script to run after each reboot and make the soundcard available
 i2s_driver_script=$RPI_SETUP_DIR/resources/load_i2s_driver.sh
 rm -f $i2s_driver_script
+if [ $# -eq 1 ] && [ $1 = "codama" ] ; then
+    sudo cp $RPI_SETUP_DIR/loader/i2s_slave/loader.ko /lib/modules/`uname -r`/kernel/sound/drivers/
+    sudo sed -i -e '$ a loader' /etc/modules-load.d/modules.conf
+else
 echo "cd $RPI_SETUP_DIR"                          >> $i2s_driver_script
 if [ $# -ge 1 ] && [ $1 = "xvf3510" ] ; then
 	echo "sudo insmod loader/i2s_master/loader.ko"               >> $i2s_driver_script
 else
 	echo "sudo insmod loader/i2s_slave/loader.ko"               >> $i2s_driver_script
+fi
 fi
 
 
@@ -97,6 +104,9 @@ sudo sed -i -e 's/#\?dtparam=i2c_arm=on/dtparam=i2c_arm=off/' /boot/config.txt
 if ! grep -q "i2c-bcm2708" /etc/modules-load.d/modules.conf; then
   sudo sh -c 'echo i2c-bcm2708 >> /etc/modules-load.d/modules.conf'
 fi
+if ! grep -q "i2c-dev" /etc/modules-load.d/modules.conf; then
+  sudo sh -c 'echo i2c-dev >> /etc/modules-load.d/modules.conf'
+fi
 if ! grep -q "options i2c-bcm2708 combined=1" /etc/modprobe.d/i2c.conf; then
   sudo sh -c 'echo "options i2c-bcm2708 combined=1" >> /etc/modprobe.d/i2c.conf'
 fi
@@ -111,27 +121,38 @@ popd > /dev/null
 # Create script to insert module into the kernel
 i2c_driver_script=$RPI_SETUP_DIR/resources/load_i2c_gpio_driver.sh
 rm -f $i2c_driver_script
-echo "cd $RPI_SETUP_DIR/i2c-gpio-param"                                            >> $i2c_driver_script
-echo "# Load the i2c bit banged driver"                                            >> $i2c_driver_script
-echo "sudo insmod i2c-gpio-param.ko"                                               >> $i2c_driver_script
-echo "# Instantiate a driver at bus id=1 on same pins as hw i2c with 1sec timeout" >> $i2c_driver_script
-echo "sudo sh -c 'echo "1 2 3 5 100 0 0 0" > /sys/class/i2c-gpio/add_bus'"         >> $i2c_driver_script
-echo "# Remove the default i2c-gpio instance"                                      >> $i2c_driver_script
-echo "sudo sh -c 'echo 7 > /sys/class/i2c-gpio/remove_bus'"                        >> $i2c_driver_script
+if [ $# -eq 1 ] && [ $1 = "codama" ] ; then
+    sudo cp $RPI_SETUP_DIR/i2c-gpio-param/i2c-gpio-param.ko /lib/modules/`uname -r`/kernel/drivers/i2c/
+    sudo depmod -ae
+    sudo sed -i -e '$ a i2c-gpio-param' /etc/modules-load.d/modules.conf
+    sudo sed -i -e '$ a options i2c-gpio-param busid=1 sda=2 scl=3 udelay=5 timeout=100 sda_od=0 scl_od=0 scl_oo=0' /etc/modprobe.d/i2c.conf
+else
+    echo "cd $RPI_SETUP_DIR/i2c-gpio-param"                                            >> $i2c_driver_script
+    echo "# Load the i2c bit banged driver"                                            >> $i2c_driver_script
+    echo "sudo insmod i2c-gpio-param.ko"                                               >> $i2c_driver_script
+    echo "# Instantiate a driver at bus id=1 on same pins as hw i2c with 1sec timeout" >> $i2c_driver_script
+    echo "sudo sh -c 'echo "1 2 3 5 100 0 0 0" > /sys/class/i2c-gpio/add_bus'"         >> $i2c_driver_script
+    echo "# Remove the default i2c-gpio instance"                                      >> $i2c_driver_script
+    echo "sudo sh -c 'echo 7 > /sys/class/i2c-gpio/remove_bus'"                        >> $i2c_driver_script
 
-echo "# Run Alsa at startup so that alsamixer configures"                          >> $i2c_driver_script
-echo "arecord -d 1 > /dev/null 2>&1"                                               >> $i2c_driver_script
-echo "aplay dummy > /dev/null 2>&1"                                                >> $i2c_driver_script
+    echo "# Run Alsa at startup so that alsamixer configures"                          >> $i2c_driver_script
+    echo "arecord -d 1 > /dev/null 2>&1"                                               >> $i2c_driver_script
+    echo "aplay dummy > /dev/null 2>&1"                                                >> $i2c_driver_script
+fi
 
 
 # Setup the crontab to restart I2S/I2C at reboot
-rm -f $RPI_SETUP_DIR/resources/crontab
-echo "@reboot sh $i2s_driver_script"  >> $RPI_SETUP_DIR/resources/crontab
-echo "@reboot sh $i2c_driver_script"  >> $RPI_SETUP_DIR/resources/crontab
-if [ $# -ge 1 ] && [ $1 = "xvf3510" ] ; then
-    echo "@reboot sh $i2s_clk_dac_script" >> $RPI_SETUP_DIR/resources/crontab
+if [ $# -eq 1 ] && [ $1 = "codama" ] ; then
+  echo "non generate crontab"
+else
+  rm -f $RPI_SETUP_DIR/resources/crontab
+  echo "@reboot sh $i2s_driver_script"  >> $RPI_SETUP_DIR/resources/crontab
+  echo "@reboot sh $i2c_driver_script"  >> $RPI_SETUP_DIR/resources/crontab
+  if [ $# -ge 1 ] && [ $1 = "xvf3510" ] ; then
+      echo "@reboot sh $i2s_clk_dac_script" >> $RPI_SETUP_DIR/resources/crontab
+  fi
+  crontab $RPI_SETUP_DIR/resources/crontab
 fi
-crontab $RPI_SETUP_DIR/resources/crontab
 
 echo "To enable I2S, this Raspberry Pi must be rebooted."
